create BitcointalkParse class
