+++ crypto-scraper/app/Console/Commands/CryptoParser.php
@@ -141,42 +141,4 @@ class CryptoParser extends Command {
         return $parsedUrl["scheme"] . "://" . $parsedUrl["host"];
     }
 
-    protected function getNextPage(string $url): ?string {
-        $crawler = $this->getPageCrawler($url);
-        $node = $crawler->filterXPath('//span[@class="prevnext"][2]/a/@href')->getNode(0);
-
-        if ($node) {
-            $nextPage = $node->nodeValue;
-            $this->printVerbose3("<fg=blue>Next page: " . $nextPage ."</>");
-            return $nextPage;
-        }
-        
-        return null;
-    }
-
-    /**
-     * @param string $url "board|topic|action=profile"
-     * @param string $pageType
-     * @return array
-     */
-    protected function getLinksFromPage(string $url, string $pageType): array {
-        $crawler = $this->getPageCrawler($url);
-        $allLinks = $crawler->filterXPath('//a[contains(@href,"https://bitcointalk.org/index.php?' . $pageType. '")]/@href')->each(function (Crawler $node) {
-            return $node->getNode(0)->nodeValue;
-        });
-        return array_unique($allLinks);
-    }
-
-    protected function getMaxPage(string $url): ?string {
-        $crawler = $this->getPageCrawler($url);
-        $node = $crawler->filterXPath('//td/a[@class="navPages"][last()]/@href')->getNode(0);
-
-        if ($node) {
-            $nextPage = $node->nodeValue;
-            $this->printVerbose3("<fg=blue>Max page: " . $nextPage ."</>");
-            return $nextPage;
-        }
-
-        return null;
-    }
 }
